% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycookie.R
\name{cookies}
\alias{cookies}
\alias{cookies_get}
\alias{cookies_set}
\alias{cookies_remove}
\title{JavaScript cookies with Shiny}
\usage{
cookies_get(inputId, cookie = NULL, session = getDefaultReactiveDomain())

cookies_set(
  cookie,
  value,
  options = list(),
  session = getDefaultReactiveDomain()
)

cookies_remove(cookie, options = list(), session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The Shiny input id to be populated with the cookie value.}

\item{cookie}{The name of the cookie to access. If left \code{NULL} in
\code{cookies_get()}, the input receives a named list of all current cookies.}

\item{session}{The current \link[shiny:session]{shiny::session} object.}

\item{value}{A string value to set the specified cookie to.}

\item{options}{A list of
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies}{attributes} to
set when creating a cookie, e.g. \code{Secure} or \code{SameSite}. When removing a
cookie, the attributes must match those that were used when it was created.}
}
\description{
Get, set and remove JavaScript cookies from your Shiny server function.
Make sure to call \code{useShinycookie()} in your UI to make the functions work.
}
\seealso{
\href{https://github.com/js-cookie/js-cookie}{JavaScript Cookie}, the
library that powers the JavaScript side of this package.
}
