% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initShinyCookie.R
\name{initShinyCookie}
\alias{initShinyCookie}
\title{Initialize shinyCookie in an applicatoin's UI}
\usage{
initShinyCookie(id, timeout = 500)
}
\arguments{
\item{id}{The identifier that will be used to reference cookies from the server as input$id}

\item{timeout}{The length of time (in milliseconds) to poll for changes to cookies}
}
\description{
This initializes the shinyCookie object in the
UI of the Shiny application. It should be placed
within the UI of the Shiny application before any
other `shinyCookie` functions are called.
}
