% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCookie.R
\name{removeCookie}
\alias{removeCookie}
\title{Remove Cookie}
\usage{
removeCookie(name, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{name}{The name of the cookie to remove}

\item{...}{name=value pairs used as attributes to specify the cookie to be deleted}

\item{session}{The shiny session}
}
\description{
Remove Cookie
}
