% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycookie-package.R
\docType{package}
\name{shinycookie-package}
\alias{shinycookie}
\alias{shinycookie-package}
\title{shinycookie: Interact with JavaScript Cookies in Shiny}
\description{
Get, set and remove JavaScript cookies in a Shiny app.
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
