% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCookie.R
\name{updateCookie}
\alias{updateCookie}
\title{Update Cookie}
\usage{
updateCookie(session, name, value)
}
\arguments{
\item{session}{The Shiny session variable}

\item{name}{The name of the cookie to update}

\item{value}{The value to assign to the cookie}
}
\description{
Sets / updates a cookie's value in the browser.
Serialization is handled by default (in `jsonlite`?).
}
