% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycookie.R
\name{useShinycookie}
\alias{useShinycookie}
\title{Set up a Shiny app to use shinycookie}
\usage{
useShinycookie()
}
\description{
Call this at least once in your UI code to make the JavaScript work. See
\link{cookies} for the enabled functionality.
}
