% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initShinyCookie.R
\name{initShinyCookie}
\alias{initShinyCookie}
\title{Initialize shinyCookie in an applicatoin's UI}
\usage{
initShinyCookie(id, namespace, timeout = 500)
}
\description{
Initialize shinyCookie in an applicatoin's UI
}
