% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCookie.R
\docType{package}
\name{shinyCookie-package}
\alias{shinyCookie}
\alias{shinyCookie-package}
\title{shinyCookie: Cookies in Shiny}
\description{
A package that allows reading and writing cookies in Shiny apps without writing custom JavaScript
}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
