% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCookie.R
\docType{package}
\name{shinyCookie-package}
\alias{shinyCookie}
\alias{shinyCookie-package}
\title{shinyCookie: Cookies in Shiny}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

}
