% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycookie.R
\docType{package}
\name{shinycookie-package}
\alias{shinycookie}
\alias{shinycookie-package}
\title{shinycookie: Cookies in Shiny}
\description{
A package that allows reading and writing cookies in Shiny apps without writing custom JavaScript. Uses the js-cookie library
}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Klaus Hartl (js-cookie) [copyright holder]
  \item Fagner Brack (js-cookie) [copyright holder]
}

}
