% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCookie.R
\name{updateCookie}
\alias{updateCookie}
\title{Update Cookie}
\usage{
updateCookie(session, ...)
}
\arguments{
\item{session}{The Shiny session variable}

\item{...}{name=value pairs to be used in setting cookie values}
}
\description{
Sets / updates a cookie's value in the browser.
Serialization is handled by default (in `jsonlite`?).
}
